install react-native-navigation
This will initialize the react-native-navigation package. Notice that there also exists alternative package for this called @react-navigation/native which is the standard package for implementing navigation stuff in React. The package we are using is created by Wix and uses platform native navigation tools and should therefore have better performance. This is important for our project as we are dealing with mapping stuff which may get quite intensive in terms of performance requirements.